+++ CHANGELOG.md
@@ -6,7 +6,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 
 # Unreleased
 ### Changed
-
+- Refactor FaviconDataAccess to use Nextclouds IClientService (#3671)
 ### Fixed
 
 # Releases